Combining retryable into one class with wrapper functions.
